requirements.txt
python-hcl2 ~= 0.3.0